Move internal spectrum processing to private module
- Move annotate_spectrum, targets_from_annotations, and
  load_and_match_spectra from spectrum_input.py to new
  _spectrum_processing.py module
- spectrum_input.py retains only the public read_spectrum_file function
- Remove unused type: ignore comments for spectrum_utils imports
  (now resolved when installed, suppressed at assignment for optional dep)
- Fix spectrum_utils type ignores to target the actual issue
  (None assignment, not unresolved import)
